\name{Internal}
\alias{Red}
\alias{Know}
\alias{DontKnow}
\alias{Unlock}
\alias{Next}
\alias{Esc}
\alias{Skip}
\title{ Internal functions for unknownR }
\description{
  Internal functions for unknownR
}
\details{
  Not for use by users. Without them exported, the tcltk 'after' function doesn't find them.
}




